[wip] separate models
