+++ backend_project/backend/api/models/__init__.py
@@ -1,0 +1,40 @@
+from .registrant import Registrant
+
+from .agency import Agency
+from .agency_jp import Agency_jp
+from .calendar import Calendar
+from .calendar_date import Calendar_Dates
+from .fare_attributes import Fare_Attributes
+from .feed_info import Feed_Info
+from .office_jp import Office_jp
+from .routes import Routes
+from .routes_jp import Routes_jp
+from .shapes import Shapes
+#from .translation import Translations
+from .trips import Trips
+from .frequencies import Frequencies
+from .stop_times import Stop_Times
+from .stops import Stops
+from .fare_rules import Fare_Rules
+from .transfers import Transfers
+
+__all__ = [
+  'Registrant',
+  'Agency',
+  'Agencyjp',
+  'Calendar',
+  'Calendar_Dates',
+  'Fare_Attributes',
+  'Fare_Rules',
+  'Feed_Info',
+  'Frequencies',
+  'Office_jp',
+  'Routes',
+  'Routes_jp',
+  'Shapes',
+  'Stops',
+  'Stop_Times',
+  'Transfers',
+#  'Translations',
+  'Trips',
+]
