+++ backend_project/backend/api/models/translations.py
@@ -1,0 +1,9 @@
+#class Translations(models.Model):
+#    table_name = models.CharField(max_length=255)
+#    field_name = models.CharField(max_length=255)
+#    language = models.CharField(max_length=255)
+#    translation = models.CharField(max_length=255)
+#    trans_id = models.CharField(max_length=255)
+#    lang = models.CharField(max_length=255)
+#    record_id = models.CharField(max_length=255)
+#    record_sub_id = models.CharField(max_length=255)
